+++ pyqgisserver/zeromq/utils.py
@@ -1,7 +1,11 @@
 
 from ..config import confservice
 
-def _get_ipc( name ) -> str:
-    ipc_path = confservice['zmq']['ipcpath']
-    return f"ipc://{ipc_path}/{name}"
+def _get_ipc( name: str ) -> str:
+    ipc_path = confservice.get('zmq','ipcpath', fallback=None)
+    if ipc_path:
+        return f"ipc://{ipc_path}/{name}"
+    else:
+        # Get alternate tcp configuration 
+        return confservice.get('zmq', f'{name}_addr')