+++ tests/run_proxy.sh
@@ -16,13 +16,16 @@ $PIP_INSTALL -q --prefer-binary -r requirements.txt
 
 $PIP install -e ./
 
+export QGIS_DISABLE_MESSAGE_HOOKS=1
+export QGIS_NO_OVERRIDE_IMPORT=1
+
 if [ -e /amqp_src ]; then
     $PIP install -e /amqp_src/
 fi
 
 # Run the server locally
 echo "Running server..."
-$VENV_PATH/bin/qgisserver -b 0.0.0.0 -p 8080 --proxy -j1
+exec $VENV_PATH/bin/qgisserver -b 0.0.0.0 -p 8080 --proxy