+++ tests/run_worker.sh
@@ -16,16 +16,16 @@ $PIP_INSTALL -q --prefer-binary -r requirements.txt
 
 $PIP install -e ./
 
+export QGIS_DISABLE_MESSAGE_HOOKS=1
+export QGIS_NO_OVERRIDE_IMPORT=1
+
 if [ -e /amqp_src ]; then
     $PIP install -e /amqp_src/
 fi
 
-export QGIS_DISABLE_MESSAGE_HOOKS=1
-export QGIS_NO_OVERRIDE_IMPORT=1
-
 # Run the server locally
 echo "Running worker..."
-$VENV_PATH/bin/qgisserver-worker --rootdir=$(pwd)/tests/data --host=$ROUTER_HOST
+exec $VENV_PATH/bin/qgisserver-worker --proxy-host=$QGSRV_ZMQ_HOSTADDR